requirements.txt
mcp>=0.9.0
python-docx>=1.1.0
pdfplumber>=0.9.0
python-pptx>=0.6.23
lxml>=4.9.0
openpyxl>=3.1.0
pandas>=2.0.0
